+++ coprocessorInt/src/LMS7002M_lib/LMS7002M_rxtsp.c
@@ -56,7 +56,7 @@ void LMS7002M_rxtsp_set_decim(LMS7002M_t *self, const LMS7002M_chan_t channel, c
     LMS7002M_regs_spi_write(self, 0x0403);
 }
 
-void LMS7002M_rxtsp_set_freq(LMS7002M_t *self, const LMS7002M_chan_t channel, const double freqRel)
+void LMS7002M_rxtsp_set_freq(LMS7002M_t *self, const LMS7002M_chan_t channel, const float freqRel)
 {
     LMS7002M_set_mac_ch(self, channel);
     self->regs->reg_0x040c_cmix_byp = (freqRel==0.0)?1:0;
@@ -157,8 +157,8 @@ void LMS7002M_rxtsp_set_dc_correction(
 void LMS7002M_rxtsp_set_iq_correction(
     LMS7002M_t *self,
     const LMS7002M_chan_t channel,
-    const double phase,
-    const double gain)
+    const float phase,
+    const float gain)
 {
     LMS7002M_set_mac_ch(self, channel);
